src/test/java/dev/dobicinaitis/advent/Day7Test.java
package dev.dobicinaitis.advent;

import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.assertEquals;

class Day7Test {

    private final Puzzle puzzle = new Day7();

    @Test
    void firstPuzzleExampleTest() {
        assertEquals(6440, puzzle.getFirstPuzzleSolution(puzzle.exampleInputFilename));
    }

    @Test
    void firstPuzzleMyInputTest() {
        assertEquals(251136060, puzzle.getFirstPuzzleSolution(puzzle.myInputFilename));
    }

    @Test
    void secondPuzzleExampleTest() {
        assertEquals(5905, puzzle.getSecondPuzzleSolution(puzzle.exampleInputFilename));
    }

    @Test
    void secondPuzzleMyInputTest() {
        assertEquals(249400220, puzzle.getSecondPuzzleSolution(puzzle.myInputFilename));
    }
}